Update sample.js
Uint32Array => Uint8Array

+++ poiemaweb_lecture/js/9.lottery/sample.js
@@ -1,5 +1,5 @@
 var rng = window.crypto || window.msCrypto
-var array = new Uint32Array(6);
+var array = new Uint8Array(6);
 
 window.crypto.getRandomValues(array)
 console.log(array.map(num => (num % 45) + 1).sort().join(','));
